refactor the Link class
